pyproject.toml
[project]
name = "viinfographicvqa"
version = "0.1.0"
description = "Vietnamese Infographic Visual Question Answering - Inference and Fine-tuning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["vqa", "vision-language", "infographic", "vietnamese", "multimodal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML/DL frameworks
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.44.0",
    "accelerate>=0.33.0",
    # Vision-Language Models
    "qwen-vl-utils>=0.0.11",
    "pillow>=10.0.0",
    # Fine-tuning
    "trl>=0.9.6",
    "peft>=0.11.1",
    "datasets>=2.19.0",
    "bitsandbytes>=0.43.0; platform_system != 'Darwin'",
    # Evaluation metrics
    "python-Levenshtein>=0.20.0",
    "regex",
    # Utilities
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
wandb = [
    "wandb>=0.17.0",
]

[project.scripts]
ft-vlm-train = "ft_vlm.fine_tuning.train:cli_main"

[project.urls]
Homepage = "https://github.com/your-org/ViInfographicVQA"
Repository = "https://github.com/your-org/ViInfographicVQA"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "ft-vlm/src/ft_vlm"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "ft-vlm/src/",
    "ft-vlm/configs/",
    "data/*.json",
]

[tool.hatch.build.targets.wheel.sources]
"ft-vlm/src" = ""

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src", "ft_vlm"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests", "ft-vlm/tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
